Move general Layer types to their own file
Removing the Layer types from the 'forest' file since
they are not specific to forest layers, and are used
for all types of layers. Exporting/importing general
layer types from the 'forest' layers file can be
confusing, thus the change.
