+++ src/meta/geo/index.ts
@@ -1,12 +1,11 @@
 export type { Bounds } from './bounds'
 export type { BurnedAreasUIOptions } from './burnedAreaSource'
 export { BurnedAreaKey, burnedAreaLayersMetadata } from './burnedAreaSource'
-export type { LayerConfig } from './forest'
-export type { LayerSource } from './forest'
 export { agreementPalette, ForestKey, forestLayersMetadata, getRecipeAgreementAreaProperty } from './forest'
 export type { ForestEstimations, ForestEstimationsData } from './forestEstimations'
 export { ExtraEstimation, extraEstimationsMetadata } from './forestEstimations'
 export type { GeoStatisticsState } from './geoStatistics'
+export type { LayerConfig, LayerSource } from './layer'
 export type { MosaicOptions, MosaicSource } from './mosaic'
 export { ProtectedAreaKey, protectedAreaLayersMetadata } from './protectedAreaSource'
 export type { BurnedAreasOptions, ForestOptions, MapPanel, ProtectedAreasOptions } from './ui'
